Ajax(vitto)

+++ src/main/java/com/unical/unitransport/controller/page/utente/GestisciRitiro.java
@@ -63,16 +63,16 @@ public class GestisciRitiro {
 	public String aggiornaRitiro(HttpServletRequest req, String luogoRitiroNuovo) {
 		HttpSession session = req.getSession(true);
 		
+		Shipment spedizione = ShipmentsDAO.getByTrackingNumber((String) session.getAttribute("codiceDaAggiornare"));
+
+		
 		if (session.getAttribute("codiceDaAggiornare")!=null) {			
-			Shipment spedizione = ShipmentsDAO.getByTrackingNumber((String) session.getAttribute("codiceDaAggiornare"));
-			if (spedizione.getStatus() >= 1)
+			if (spedizione.getStatus() >= 1) {
+				session.setAttribute("erroreGenerico", "LA SPEDIZIONE NON PUO' ESSERE MODIFICATA");
+				session.setAttribute("erroreGenerico_p", "La spedizione è in stato troppo avanzato");
 				return "erroreGenerico";
-			
-			ShipmentsDAO.updateRitiro(spedizione, spedizione.getStatus(), luogoRitiroNuovo);
-			if (spedizione.getLastLocation().equals(spedizione.getSenderLocation())) {
-				spedizione.setLastLocation(luogoRitiroNuovo);
 			}
-			spedizione.setSenderLocation(luogoRitiroNuovo);
+			
 
 		
 
@@ -84,14 +84,21 @@ public class GestisciRitiro {
 				}
 			
 			if(utenteValido) {
+				ShipmentsDAO.updateRitiro(spedizione, spedizione.getStatus(), luogoRitiroNuovo);
+					if (spedizione.getLastLocation().equals(spedizione.getSenderLocation())) {
+						spedizione.setLastLocation(luogoRitiroNuovo);
+					}
+				spedizione.setSenderLocation(luogoRitiroNuovo);
+
 				session.setAttribute("validoGenerico", "LA SPEDIZIONE È STATA CORRETTAMENTE AGGIORNATA");
 				session.setAttribute("validoGenerico_p", null);
+				session.setAttribute("codiceDaAggiornare", null);
 				return "validoGenerico";
 			}
 		}
 
-		session.setAttribute("erroreGenerico", "LA SPEDIZIONE NON APPARTIENE ALL'UTENTE CHE NE FA RICHIESTA");
-		session.setAttribute("erroreGenerico_p", null);
+			session.setAttribute("erroreGenerico", "LA SPEDIZIONE NON APPARTIENE ALL'UTENTE CHE NE FA RICHIESTA");
+			session.setAttribute("erroreGenerico_p", null);
 			return "erroreGenerico";
 		
 	}
@@ -101,25 +108,47 @@ public class GestisciRitiro {
 	public String controllaTrackingElimina(HttpServletRequest req, String trackingModifica) {
 		HttpSession session = req.getSession(true);
 		
-		session.setAttribute("codiceDaAggiornare", null);
 
 		if (ShipmentsDAO.getByTrackingNumber(trackingModifica)!=null) {
-			ShipmentsDAO.remove(ShipmentsDAO.getByTrackingNumber(trackingModifica));	
+			List<String> spedizioni = ShipmentsSenderReceiverDAO.getAllString((String)req.getSession().getAttribute("email"));
+			if (ShipmentsDAO.getByTrackingNumber(trackingModifica).getStatus() >= 1) {
+				session.setAttribute("erroreGenerico", "LA SPEDIZIONE NON PUO' ESSERE MODIFICATA");
+				session.setAttribute("erroreGenerico_p", "La spedizione è in stato troppo avanzato");
+				return "erroreGenerico";
+			}
+			
+			boolean utenteValido = false;
+				for (String spedizioneUtenteCorrente: spedizioni) {
+					if (spedizioneUtenteCorrente.equals( trackingModifica  ))
+						utenteValido=true;
+					}
+			
+				if(utenteValido) {
+					ShipmentsDAO.remove(ShipmentsDAO.getByTrackingNumber(trackingModifica));	
+					session.setAttribute("codiceDaAggiornare", null);
+					session.setAttribute("validoGenerico", "LA SPEDIZIONE E' STATA ELIMINATA CON SUCCESSO");
+					session.setAttribute("validoGenerico_p", null);
+					return "validoGenerico";
+				} else {
 			session.setAttribute("codiceDaAggiornare", null);
-			session.setAttribute("validoGenerico", "LA SPEDIZIONE E' STATA ELIMINATA CON SUCCESSO");
-			session.setAttribute("validoGenerico_p", null);
-			return "validoGenerico";
-		} else {
-			session.setAttribute("codiceDaAggiornare", null);
-			session.setAttribute("erroreGenerico", "LA SPEDIZIONE NON ESISTE");
+			session.setAttribute("erroreGenerico", "LA SPEDIZIONE NON TI APPARTIENE ");
+			session.setAttribute("erroreGenerico_p", null);
 			return "erroreGenerico";
+				}
 		}
+		
+		return null;
+		
+
+		
+		
+	}
 
 		
 		
 			
 		
-	}
+	
 	
 
 	
@@ -149,12 +178,12 @@ public class GestisciRitiro {
 		HttpSession session = req.getSession(true);
 		
 		if (session.getAttribute("codiceDaAggiornare")!=null) {
-			Shipment spedizione = ShipmentsDAO.getByTrackingNumber((String) session.getAttribute("codiceDaAggiornare"));
-			ShipmentsDAO.updateLuogoConsegna(spedizione, spedizione.getStatus(), luogoConsegnaNuovo);
-			spedizione.setReceiverLocation(luogoConsegnaNuovo);
-		
-
 			List<String> spedizioni = ShipmentsSenderReceiverDAO.getAllString((String)req.getSession().getAttribute("email"));
+				if (ShipmentsDAO.getByTrackingNumber((String) session.getAttribute("codiceDaAggiornare")).getStatus() >= 1) {
+					session.setAttribute("erroreGenerico", "LA SPEDIZIONE NON PUO' ESSERE MODIFICATA");
+					session.setAttribute("erroreGenerico_p", "La spedizione è in stato troppo avanzato");
+					return "erroreGenerico";
+				}
 			boolean utenteValido = false;
 			for (String spedizioneUtenteCorrente: spedizioni) {
 				if (spedizioneUtenteCorrente.equals( (String) session.getAttribute("codiceDaAggiornare")))
@@ -162,6 +191,9 @@ public class GestisciRitiro {
 				}
 			
 			if(utenteValido) {
+				Shipment spedizione = ShipmentsDAO.getByTrackingNumber((String) session.getAttribute("codiceDaAggiornare"));
+				ShipmentsDAO.updateLuogoConsegna(spedizione, spedizione.getStatus(), luogoConsegnaNuovo);
+				spedizione.setReceiverLocation(luogoConsegnaNuovo);
 				session.setAttribute("validoGenerico", "LA SPEDIZIONE È STATA CORRETTAMENTE AGGIORNATA");
 				session.setAttribute("validoGenerico_p", null);
 				return "validoGenerico";
